Refactor controllers
