+++ controllers/index.js
@@ -1,6 +1,6 @@
 const router = require("express").Router();
 
-const htmlRoutes = require("./html");
+const htmlRoutes = require("./htmlRoutes");
 const apiRoutes = require("./api");
 
 router.use("/", htmlRoutes);
